Switching StreamEvent to use the classification pattern we use for polymorphic types.
